requirements.txt
h5py
tqdm
mujoco
pyyaml
rospkg
einops
mediapy
pexpect
ipython
packaging
matplotlib
pyquaternion
opencv-python
opencv-contrib-python
dynamixel-sdk
isort
black
scipy
open3d
lightning>=2.0.0
hydra-colorlog==1.2.0
hydra-core==1.3.2
hydra-optuna-sweeper==1.2.0
rootutils
rich
pynput
imageio[ffmpeg]
gdown
timm==0.9.10
